Refactor and remove dead code * Move code to appropriate packages     - agent websocket     - console log websocket * Remove dead code
