+++ server/src/com/thoughtworks/go/server/service/BuildAssignmentService.java
@@ -27,8 +27,8 @@ import com.thoughtworks.go.server.domain.BuildComposer;
 import com.thoughtworks.go.server.materials.StaleMaterialsOnBuildCause;
 import com.thoughtworks.go.server.service.builders.BuilderFactory;
 import com.thoughtworks.go.server.transaction.TransactionTemplate;
-import com.thoughtworks.go.server.websocket.Agent;
-import com.thoughtworks.go.server.websocket.AgentRemoteHandler;
+import com.thoughtworks.go.server.websocket.agent.Agent;
+import com.thoughtworks.go.server.websocket.agent.AgentRemoteHandler;
 import com.thoughtworks.go.util.TimeProvider;
 import com.thoughtworks.go.util.URLService;
 import com.thoughtworks.go.websocket.Action;
