+++ server/src/com/thoughtworks/go/server/websocket/agent/AgentRemoteServlet.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 ThoughtWorks, Inc.
+ * Copyright 2017 ThoughtWorks, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.thoughtworks.go.server.websocket;
+package com.thoughtworks.go.server.websocket.agent;
 
 import com.thoughtworks.go.util.SystemEnvironment;
 import org.eclipse.jetty.websocket.servlet.WebSocketServlet;
