+++ server/src/com/thoughtworks/go/server/websocket/consolelog/ConsoleLogSocket.java
@@ -14,9 +14,12 @@
  * limitations under the License.
  */
 
-package com.thoughtworks.go.server.websocket;
+package com.thoughtworks.go.server.websocket.consolelog;
 
 import com.thoughtworks.go.domain.JobIdentifier;
+import com.thoughtworks.go.server.websocket.SocketEndpoint;
+import com.thoughtworks.go.server.websocket.SocketHealthService;
+import com.thoughtworks.go.server.websocket.WebsocketMessages;
 import org.apache.http.NameValuePair;
 import org.apache.http.client.utils.URLEncodedUtils;
 import org.eclipse.jetty.websocket.api.Session;
