+++ server/test/unit/com/thoughtworks/go/server/websocket/ConsoleLogSenderTest.java
@@ -23,6 +23,7 @@ import com.thoughtworks.go.domain.JobInstance;
 import com.thoughtworks.go.domain.exception.IllegalArtifactLocationException;
 import com.thoughtworks.go.server.dao.JobInstanceDao;
 import com.thoughtworks.go.server.service.ConsoleService;
+import com.thoughtworks.go.server.websocket.consolelog.ConsoleLogSender;
 import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang.RandomStringUtils;
 import org.junit.Before;
